tunnelrender: CB1 fix P/Q compile error [1/1]
PD#SWPL-74643

Problem:
fix P/Q compile error

Solution:
fix P/Q compile error

Verify:
ohm

SourceCodec:
V3.2.216-g953bd52
953bd52 tunnelrender: CB1 fix P/Q compile error [1/1]
3abac06 mediahal: CB1 fix inputframe over-size when 4khevc [1/1]
1b2662f tunnelrender: CF2 tunnelrender code clean [1/1]
4a61db8 mediahal: CF1 Compile the 64 bit library.  [1/1]
74dce10 tsplayer: CB1 Fix 2 crashes. [1/1]
04d5c69 mediahal: CB1 Organize the license of the code  [1/1]
0fb07e3 videodec: CF0 support omx v2 on T7 andorid p. [2/3]
99e4f40 mediasync: CB1 mediasync optimization part-1 . [1/2]
c6588e2 mediasync: CB1 MediaSync_create() blocks.  [1/1]
e88511c tsplayer: CB1 VideoTunnelDisplay::Disconnect() crash [1/1]

+++ example/AmTsPlayerExample/AmTsPlayerExample.cpp
@@ -515,13 +515,13 @@ int main(int argc, char **argv)
        file.seekg(0, file.end);
        fsize = file.tellg();
        if (fsize <= 0) {
-           printf("file %s size %lld return\n", inputTsName.c_str(), fsize);
+           printf("file %s size %lld return\n", inputTsName.c_str(),(long long)fsize);
            return 0;
        }
        file.seekg(0, file.beg);
     }
     printf("file name = %s, is_open %d, size %lld, tsType %d\n",
-                inputTsName.c_str(), file.is_open(), fsize, tsType);
+                inputTsName.c_str(), file.is_open(),(long long) fsize, tsType);
 
     //am_tsplayer_handle session;
     am_tsplayer_init_params parm = {tsType, drmmode, 0, 0};
@@ -660,7 +660,7 @@ int main(int argc, char **argv)
                 state.data = (uint8_t *)malloc(state.data_len);
                 memset(state.data, 0x0, state.data_len);
                 AmTsPlayer_getState(session, &state);
-                printf("----json[%d]:\n%s\n", state.actual_len, state.data);
+                printf("----json[%d]:\n%s\n", (int)state.actual_len, state.data);
                 free(state.data);
             }
         }
